app/requirements.txt
alembic       ==    1.11.1
fastapi      ==     0.96.0
fastapi-jwt-auth == 0.5.0
psycopg2-binary ==  2.9.6
python-dotenv  ==   1.0.0
SQLAlchemy   ==     2.0.15
uvicorn        ==   0.22.0
Werkzeug     ==     2.3.6
